Derived builders only implement methods that differ
