+++ backend/antibiootit/src/main/java/fi/tuni/koodimankelit/antibiootit/builder/AntibioticTreatmentBuilder.java
@@ -3,11 +3,15 @@ package fi.tuni.koodimankelit.antibiootit.builder;
 import fi.tuni.koodimankelit.antibiootit.database.data.Antibiotic;
 import fi.tuni.koodimankelit.antibiootit.database.data.Strength;
 import fi.tuni.koodimankelit.antibiootit.models.AntibioticTreatment;
+import fi.tuni.koodimankelit.antibiootit.models.DosageFormula;
+import fi.tuni.koodimankelit.antibiootit.models.DosageResult;
+import fi.tuni.koodimankelit.antibiootit.models.Instructions;
 /**
  * Builder for antibiotic treatment. Calculates dosage based on weight
  */
 public abstract class AntibioticTreatmentBuilder {
 
+    private final Antibiotic antibiotic;
     protected final double weight;
     protected final Strength strength;
 
@@ -18,6 +22,7 @@ public abstract class AntibioticTreatmentBuilder {
      */
     public AntibioticTreatmentBuilder(Antibiotic antibiotic, double weight) {
 
+        this.antibiotic = antibiotic;
         this.weight = weight;
         this.strength = getStrength(antibiotic, weight);
     }
@@ -27,8 +32,29 @@ public abstract class AntibioticTreatmentBuilder {
      * Build antibiotic treatment object
      * @return AntibioticTreatment generated result
      */
-    public abstract AntibioticTreatment build();
+    public AntibioticTreatment build() {
 
+        // Does not exceed minimum weight on any strength
+        if(this.strength == null) {
+            // TODO change to custom exception
+            throw new RuntimeException("Selected antibiotic has no suitable strength");
+        }
+
+        Instructions instructions = new Instructions(antibiotic.getDays(), antibiotic.getDosesPerDay());
+
+        return new AntibioticTreatment(
+            antibiotic.getFormat(),
+            antibiotic.getInfo(),
+            antibiotic.getAntibiotic(),
+            instructions,
+            buildFormula(),
+            buildResult()
+        );
+    }
+
+    protected abstract DosageFormula buildFormula();
+
+    protected abstract DosageResult buildResult();
     
     /** 
      * Return antibiotic's strength
