+++ backend/antibiootit/src/main/java/fi/tuni/koodimankelit/antibiootit/builder/MixtureBuilder.java
@@ -1,10 +1,8 @@
 package fi.tuni.koodimankelit.antibiootit.builder;
 
 import fi.tuni.koodimankelit.antibiootit.database.data.Mixture;
-import fi.tuni.koodimankelit.antibiootit.models.AntibioticTreatment;
 import fi.tuni.koodimankelit.antibiootit.models.DosageFormula;
 import fi.tuni.koodimankelit.antibiootit.models.DosageResult;
-import fi.tuni.koodimankelit.antibiootit.models.Instructions;
 import fi.tuni.koodimankelit.antibiootit.models.Measurement;
 
 public class MixtureBuilder extends AntibioticTreatmentBuilder {
@@ -18,32 +16,18 @@ public class MixtureBuilder extends AntibioticTreatmentBuilder {
     }
 
     @Override
-    public AntibioticTreatment build() {
-
-        // Does not exceed minimum weight on any strength
-        if(this.strength == null) {
-            // TODO change to custom exception
-            throw new RuntimeException("Selected antibiotic has no suitable strength");
-        }
-
-        Instructions instructions = new Instructions(antibiotic.getDays(), antibiotic.getDosesPerDay());
-        DosageFormula dosageFormula = new DosageFormula(
+    protected DosageFormula buildFormula() {
+        return new DosageFormula(
             new Measurement(strength.getUnit(), strength.getValue()),
             new Measurement(antibiotic.getDosagePerWeightPerDayUnit(), antibiotic.getDosagePerWeightPerDay())
         );
-        DosageResult dosageResult = new DosageResult(
+    }
+
+    @Override
+    protected DosageResult buildResult() {
+        return new DosageResult(
             new Measurement(antibiotic.getResultUnit(), calculateDosageResult())
         );
-
-        return new AntibioticTreatment(
-            antibiotic.getFormat(),
-            antibiotic.getInfo(),
-            antibiotic.getAntibiotic(),
-            instructions,
-            dosageFormula,
-            dosageResult
-        );
-            
     }
 
     /**
