+++ backend/antibiootit/src/main/java/fi/tuni/koodimankelit/antibiootit/builder/TabletBuilder.java
@@ -2,10 +2,8 @@ package fi.tuni.koodimankelit.antibiootit.builder;
 
 
 import fi.tuni.koodimankelit.antibiootit.database.data.Tablet;
-import fi.tuni.koodimankelit.antibiootit.models.AntibioticTreatment;
 import fi.tuni.koodimankelit.antibiootit.models.DosageFormula;
 import fi.tuni.koodimankelit.antibiootit.models.DosageResult;
-import fi.tuni.koodimankelit.antibiootit.models.Instructions;
 import fi.tuni.koodimankelit.antibiootit.models.Measurement;
 
 public class TabletBuilder extends AntibioticTreatmentBuilder {
@@ -19,24 +17,17 @@ public class TabletBuilder extends AntibioticTreatmentBuilder {
     }
 
     @Override
-    public AntibioticTreatment build() {
-        
-        Instructions instructions = new Instructions(antibiotic.getDays(), antibiotic.getDosesPerDay());
-        DosageFormula dosageFormula = new DosageFormula(
+    protected DosageFormula buildFormula() {
+        return new DosageFormula(
             new Measurement(strength.getUnit(), strength.getValue()),
             new Measurement(antibiotic.getDosagePerDayUnit(), antibiotic.getTabletsPerDose() * strength.getValue())
         );
-        DosageResult dosageResult = new DosageResult(
-            new Measurement("kpl", antibiotic.getTabletsPerDose())
-        );
+    }
 
-        return new AntibioticTreatment(
-            antibiotic.getFormat(),
-            antibiotic.getInfo(),
-            antibiotic.getAntibiotic(),
-            instructions,
-            dosageFormula,
-            dosageResult
+    @Override
+    protected DosageResult buildResult() {
+        return new DosageResult(
+            new Measurement("kpl", antibiotic.getTabletsPerDose())
         );
     }
